requirements.txt
# RTSP Camera Dashboard Requirements

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# Video Processing
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.1.0

# HTTP Client for SmolVLM
requests==2.31.0

# Async Support
eventlet==0.33.3
python-socketio==5.10.0
